working

+++ toClass.py
@@ -5,10 +5,14 @@ import os,sys,re,json,argparse
 parser = argparse.ArgumentParser('make a python script into a class')
 
 parser.add_argument('-v', '--verbose', action='store_true',                    help='show verbose output')
+parser.add_argument('-i', '--inplace', action='store_true',                    help='replace existing file')
+parser.add_argument('-o', '--output',  action='store',                         help='output to file')
 parser.add_argument('-t', '--tab',     action='store',     default='    ',     help='indent characters')
 parser.add_argument('-c', '--clasz',   action='store',     required=True,      help='the name of the class required')
-parser.add_argument('-i', '--include', action='store',     default='^\^',      help='regex pattern to include in class <name> ():')
-parser.add_argument('-m', '--main',    action='store',     default='^\:',      help='regex pattern to include in main():')
+parser.add_argument('-a', '--attr',    action='store',     default='@',        help='attribute prefix to harvest attributes')
+parser.add_argument('-b', '--body',    action='store',     default='^',        help='method prefix fo harvest methods')
+parser.add_argument('-m', '--main',    action='store',     default=':',        help='main prefix to harvest main __init__()')
+parser.add_argument('-l', '--last',    action='store',     default='$',        help='last prefix insert main() call')
 parser.add_argument('file',            action='store',     nargs='*',          help='the file name')
 
 args = parser.parse_args()
@@ -23,15 +27,15 @@ class State(object):
 
     _instance = None
 
-    classPattern  = re.compile('%s.*'%args.include)
-    attributePattern = re.compile('%s.*'%args.include)
-    methodPattern = re.compile('^%sdef\s([^\(]+)\(([^\)]*)\)\:(.*)$'%args.include)
-    mainPattern   = re.compile('%s.*'%args.main)
-
+    classPattern  = re.compile('^\\%s.*'%args.body)
+    attrPattern   = re.compile('^\\%s.*'%args.attr)
+    mainPattern   = re.compile('^\\%s.*'%args.main)
+    lastPattern   = re.compile('^\\%s.*'%args.last)
+     
     def __new__(cls,*args,**kwargs):
         ''' singleton pattern '''
         if not cls._instance:
-            cls.instance = super(State,cls).__new__(cls,*args,**kwargs)
+            cls._instance = super(State,cls).__new__(cls,*args,**kwargs)
         return cls._instance
 
     def __init__(self):
@@ -48,7 +52,7 @@ class BeforeClass(State):
     def before(self,line,output):
         return
     def process(self,line,output):
-        if self.classPattern.match(line):
+        if self.classPattern.match(line) or self.attrPattern.match(line):
             state = InClass()
             state.before(line,output)
             return state
@@ -57,11 +61,17 @@ class BeforeClass(State):
 
 ################################################################################
 class InClass(State):
+    methodPattern = re.compile('^\\%sdef\s([^\(]+)\(([^\)]*)\)\:(.*)$'%args.body)
     def before(self,line,output):
-        output.write('class %s(object):\n'%args.clasz)
-        output.write('%s%s\n'%(args.tab,line.lstrip('^')))
+        output.write('class %s(object):\n\n'%args.clasz)
+        output.write('%s%s\n'%(args.tab,line.lstrip('[%s%s]'%(args.attr,args.body))))
         return
     def process(self,line,output):
+        isMethod = self.methodPattern.match(line)
+        if isMethod:
+            (name,params,rest) = isMethod.groups()
+            output.write('\n%sdef %s(self,%s):%s\n'%(args.tab,name,params,rest))
+            return self
         if self.mainPattern.match(line):
             state = InMain()
             state.before(line,output)
@@ -69,28 +79,31 @@ class InClass(State):
         if self.classPattern.match(line):
             output.write('%s%s\n'%(args.tab,line.lstrip('^')))
             return self
-        state = AfterAll()
-        state.before(line,output)
-        return state
+        if self.lastPattern.match(line):
+            state = AfterAll()
+            state.before(line,output)
+            return state
+        return self
 
 ################################################################################
 class InMain(State):
     def before(self,line,output):
-        output.write('%sdef __init__(self):\n'%args.tab)
-        output.write('%s%s\n'%(args.tab*2,line.lstrip(':')))
+        output.write('\n%sdef __init__(self):\n'%args.tab)
+        output.write('%s%s\n'%(args.tab*2,line.lstrip(args.main)))
         return
     def process(self,line,output):
-        output.write('%s%s\n'%(args.tab*2,line.lstrip(':')))
-        return
+        if self.lastPattern.match(line):
+            state = AfterAll()
+            state.before(line,output)
+            return state
+        output.write('%s%s\n'%(args.tab*2,line.lstrip(args.main)))
+        return self
 
 ################################################################################
 class AfterAll(State):
     def before(self,line,output):
-        output.write('''
-if __name__ == '__main__': main()
-def main():
-%sinstance=%s()
-'''%(args.tab,args.clasz))
+        output.write('\ndef main():\n%sinstance=%s()\n'%(args.tab,args.clasz))
+        output.write('\nif __name__ == \'__main__\': main()\n')
         return
     def process(self,line,output):
         output.write('%s\n'%line)
@@ -109,16 +122,21 @@ def process(input,output):
 
 def main():
     for file in args.file:
-        backup='%s.bak'
-        if os.path.isfile(backup):
-            os.unlink(backup)
-        os.rename(file,backup)
-        sys.stderr.write('%s\n'%file)
-        input=open(backup)
-        output=open(file,'w')
-        process(input,output)
-        output.close()
-        input.close()
+        if args.inplace:
+            backup='%s.bak'
+            if os.path.isfile(backup):
+                os.unlink(backup)
+            os.rename(file,backup)
+            sys.stderr.write('%s\n'%file)
+            input=open(backup)
+            output=open(file,'w')
+            process(input,output)
+            output.close()
+            input.close()
+        else:
+            input=open(file)
+            output=sys.stdout
+            process(input,output)
     return
 
 if __name__ == '__main__': main()
